cellGame slightly rearranged
