+++ rpcs3/Emu/SysCalls/Modules/cellSysutil.cpp
@@ -16,10 +16,7 @@
 #include "Emu/Audio/AudioManager.h"
 #include "Emu/FS/VFS.h"
 #include "cellMsgDialog.h"
-#include "cellGame.h"
 #include "cellSysutil.h"
-
-typedef void (CellHddGameStatCallback)(vm::ptr<CellHddGameCBResult> cbResult, vm::ptr<CellHddGameStatGet> get, vm::ptr<CellHddGameStatSet> set);
 
 extern Module cellSysutil;
 
@@ -688,76 +685,6 @@ int cellSysCacheMount(vm::ptr<CellSysCacheParam> param)
 	return CELL_SYSCACHE_RET_OK_RELAYED;
 }
 
-int cellHddGameCheck(u32 version, vm::ptr<const char> dirName, u32 errDialog, vm::ptr<CellHddGameStatCallback> funcStat, u32 container)
-{
-	cellSysutil.Warning("cellHddGameCheck(version=%d, dirName_addr=0x%x, errDialog=%d, funcStat_addr=0x%x, container=%d)",
-		version, dirName.addr(), errDialog, funcStat.addr(), container);
-
-	std::string dir = dirName.get_ptr();
-	if (dir.size() != 9)
-		return CELL_HDDGAME_ERROR_PARAM;
-
-	vm::var<CellHddGameSystemFileParam> param;
-	vm::var<CellHddGameCBResult> result;
-	vm::var<CellHddGameStatGet> get;
-	vm::var<CellHddGameStatSet> set;
-
-	get->hddFreeSizeKB = 40 * 1024 * 1024; // 40 GB, TODO: Use the free space of the computer's HDD where RPCS3 is being run.
-	get->isNewData = CELL_HDDGAME_ISNEWDATA_EXIST;
-	get->sysSizeKB = 0; // TODO
-	get->st_atime__  = 0; // TODO
-	get->st_ctime__  = 0; // TODO
-	get->st_mtime__  = 0; // TODO
-	get->sizeKB = CELL_HDDGAME_SIZEKB_NOTCALC;
-	memcpy(get->contentInfoPath, ("/dev_hdd0/game/" + dir).c_str(), CELL_HDDGAME_PATH_MAX);
-	memcpy(get->hddGamePath, ("/dev_hdd0/game/" + dir + "/USRDIR").c_str(), CELL_HDDGAME_PATH_MAX);
-
-	if (!Emu.GetVFS().ExistsDir(("/dev_hdd0/game/" + dir).c_str()))
-	{
-		get->isNewData = CELL_HDDGAME_ISNEWDATA_NODIR;
-	}
-	else
-	{
-		// TODO: Is cellHddGameCheck really responsible for writing the information in get->getParam ? (If not, delete this else)
-
-		vfsFile f(("/dev_hdd0/game/" + dir + "/PARAM.SFO").c_str());
-		PSFLoader psf(f);
-		if (!psf.Load(false)) {
-			return CELL_HDDGAME_ERROR_BROKEN;
-		}
-
-		get->getParam.parentalLevel = psf.GetInteger("PARENTAL_LEVEL");
-		get->getParam.attribute = psf.GetInteger("ATTRIBUTE");
-		get->getParam.resolution = psf.GetInteger("RESOLUTION");
-		get->getParam.soundFormat = psf.GetInteger("SOUND_FORMAT");
-		std::string title = psf.GetString("TITLE");
-		strcpy_trunc(get->getParam.title, title);
-		std::string app_ver = psf.GetString("APP_VER");
-		strcpy_trunc(get->getParam.dataVersion, app_ver);
-		strcpy_trunc(get->getParam.titleId, dir);
-
-		for (u32 i=0; i<CELL_HDDGAME_SYSP_LANGUAGE_NUM; i++) {
-			char key [16];
-			sprintf(key, "TITLE_%02d", i);
-			title = psf.GetString(key);
-			strcpy_trunc(get->getParam.titleLang[i], title);
-		}
-	}
-
-	// TODO ?
-
-	funcStat(result, get, set);
-
-	if (result->result != CELL_HDDGAME_CBRESULT_OK &&
-        result->result != CELL_HDDGAME_CBRESULT_OK_CANCEL) {
-		return CELL_HDDGAME_ERROR_CBRESULT;
-    }
-
-	// TODO ?
-
-	return CELL_OK;
-}
-
 bool bgm_playback_enabled = true;
 
 int cellSysutilEnableBgmPlayback()
@@ -835,13 +762,8 @@ int cellWebBrowserEstimate2(const vm::ptr<const CellWebBrowserConfig2> config, v
 	return CELL_OK;
 }
 
-extern int cellGameDataCheckCreate2(PPUThread& CPU, u32 version, vm::ptr<const char> dirName, u32 errDialog,
-	vm::ptr<void(vm::ptr<CellGameDataCBResult> cbResult, vm::ptr<CellGameDataStatGet> get, vm::ptr<CellGameDataStatSet> set)> funcStat, u32 container);
-
-extern int cellGameDataCheckCreate(PPUThread& CPU, u32 version, vm::ptr<const char> dirName, u32 errDialog,
-	vm::ptr<void(vm::ptr<CellGameDataCBResult> cbResult, vm::ptr<CellGameDataStatGet> get, vm::ptr<CellGameDataStatSet> set)> funcStat, u32 container);
-
 extern void cellSysutil_SaveData_init();
+extern void cellSysutil_GameData_init();
 
 Module cellSysutil("cellSysutil", []()
 {
@@ -850,6 +772,9 @@ Module cellSysutil("cellSysutil", []()
 		v.func.set(0);
 		v.arg.set(0);
 	}
+
+	cellSysutil_SaveData_init(); // cellSaveData functions
+	cellSysutil_GameData_init(); // cellGameData, cellHddGame functions
 
 	REG_FUNC(cellSysutil, cellSysutilGetSystemParamInt);
 	REG_FUNC(cellSysutil, cellSysutilGetSystemParamString);
@@ -893,22 +818,8 @@ Module cellSysutil("cellSysutil", []()
 	REG_FUNC(cellSysutil, cellSysCacheMount);
 	REG_FUNC(cellSysutil, cellSysCacheClear);
 
-	REG_FUNC(cellSysutil, cellHddGameCheck);
-	//REG_FUNC(cellSysutil, cellHddGameCheck2);
-	//REG_FUNC(cellSysutil, cellHddGameGetSizeKB);
-	//REG_FUNC(cellSysutil, cellHddGameSetSystemVer);
-	//REG_FUNC(cellSysutil, cellHddGameExitBroken);
-
 	//REG_FUNC(cellSysutil, cellSysutilRegisterCallbackDispatcher);
 	//REG_FUNC(cellSysutil, cellSysutilPacketWrite);
-	//REG_FUNC(cellSysutil, doc.write);
-	//REG_FUNC(cellSysutil, packet_read);
-
-	// cellSaveData functions
-	cellSysutil_SaveData_init();
 
 	REG_FUNC(cellSysutil, cellWebBrowserEstimate2);
-
-	REG_FUNC(cellSysutil, cellGameDataCheckCreate);
-	REG_FUNC(cellSysutil, cellGameDataCheckCreate2);
 });
